+++ source/core/rootservice/src/main/kotlin/com/xayah/core/rootservice/service/RemoteRootService.kt
@@ -313,6 +313,9 @@ class RemoteRootService(private val context: Context) {
     suspend fun forceStopPackageAsUser(packageName: String, userId: Int) =
         runCatching { getService().forceStopPackageAsUser(packageName, userId) }.onFailure(onFailure)
 
+    suspend fun uninstallPackageAsUser(packageName: String, userId: Int): Boolean =
+        runCatching { getService().uninstallPackageAsUser(packageName, userId) }.onFailure(onFailure).getOrElse { false }
+
     suspend fun setApplicationEnabledSetting(packageName: String, newState: Int, flags: Int, userId: Int, callingPackage: String?) =
         runCatching { getService().setApplicationEnabledSetting(packageName, newState, flags, userId, callingPackage) }.onFailure(onFailure)